Update __init__.py
If _parse_value results in an empty string an exception is raised due to it not matching any ISO8601 format. When the exception occurs the value is set to the result of _parse_value which is again empty. This change should allow me to get the text version of the date or datetime.